Create pasteleirosController
