+++ src/routes.js
@@ -1,4 +1,4 @@
-const _ = require('lodash');
+const pasteleiroController = require('./controllers/pasteleirosController');
 
 const {
   getWebClientMiddleware,
@@ -8,15 +8,5 @@ const {
 module.exports = app => {
   app.use(getWebClientMiddleware);
 
-  app.post('/pasteleiro-do-dia', getChannelMembersMiddleware, async (req, res) => {
-    const { members } = req
-    const pasteleiro = `<@${_.sample(members)}>`
-
-    const data = {
-      response_type: 'in_channel',
-      text: `${pasteleiro} you were picked to be the Pasteleiro today. Congratz! :tada:`,
-    };
-
-    return res.json(data);
-  });
+  app.post('/pasteleiro-do-dia', getChannelMembersMiddleware, pasteleiroController.createPasteleiro);
 }
